@@ -1,54 +1,49 @@
 name: deriv_chart
 description: Chart for multiplier flutter app.
 version: 0.0.1
-author:
-homepage:
+homepage: ""
+publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.1"
 
 dependencies:
-  flutter:
+  flutter_localizations:
     sdk: flutter
-  intl: ^0.16.1
+
+  collection: ^1.15.0-nullsafety.4
+  deriv_technical_analysis:
+    git:
+      url: git@github.com:regentmarkets/flutter-deriv-packages.git
+      path: packages/deriv_technical_analysis
+      ref: dev
+
+  equatable: ^2.0.5
+  intl: ^0.18.0
+  json_annotation: ^4.8.0
+  provider: ^6.0.5
+  shared_preferences: ^2.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  deriv_lint:
+    git:
+      url: git@github.com:regentmarkets/flutter-deriv-packages.git
+      path: packages/deriv_lint
+      ref: flutter-version-3
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
+  build_runner: ^2.3.3
+  intl_utils: ^2.8.2
+  json_serializable: ^6.6.1
 
-# The following section is specific to Flutter.
 flutter:
+  assets:
+    - assets/icons/
+    - assets/icons/symbols/
 
-  # To add assets to your package, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
-  #
-  # For details regarding assets in packages, see
-  # https://flutter.dev/assets-and-images/#from-packages
-  #
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # To add custom fonts to your package, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts in packages, see
-  # https://flutter.dev/custom-fonts/#from-packages
+flutter_intl:
+  main_locale: en
+  enabled: true
+  class_name: ChartLocalization
