@@ -1,5 +1,47 @@
 library deriv_chart;
 
-export './src/models/candle.dart';
-export './src/models/chart_style.dart';
-export './src/chart.dart';
+export 'generated/l10n.dart';
+export 'src/deriv_chart/chart/chart.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/barriers/accumulators_horizontal_barrier/accumulators_entry_spot_barrier.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/barriers/barrier.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/barriers/horizontal_barrier/combined_barrier.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/barriers/horizontal_barrier/horizontal_barrier.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/barriers/horizontal_barrier/tick_indicator.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/barriers/vertical_barrier/vertical_barrier.dart';
+export 'src/deriv_chart/chart/data_visualization/annotations/chart_annotation.dart';
+export 'src/deriv_chart/chart/data_visualization/chart_series/data_series.dart';
+export 'src/deriv_chart/chart/data_visualization/chart_series/indicators_series/bollinger_bands_series.dart';
+export 'src/deriv_chart/chart/data_visualization/chart_series/indicators_series/ma_series.dart';
+export 'src/deriv_chart/chart/data_visualization/chart_series/line_series/line_series.dart';
+export 'src/deriv_chart/chart/data_visualization/chart_series/ohlc_series/candle/candle_series.dart';
+export 'src/deriv_chart/chart/data_visualization/chart_series/series.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/active_marker.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/marker.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/marker_icon_painters/accumulators_marker_icon_painter.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/marker_icon_painters/marker_icon_painter.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/marker_icon_painters/multipliers_marker_icon_painter.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/marker_icon_painters/options_marker_icon_painter.dart';
+export 'src/deriv_chart/chart/data_visualization/markers/marker_series.dart';
+export 'src/deriv_chart/chart/data_visualization/models/chart_object.dart';
+export 'src/deriv_chart/chart/helpers/paint_functions/paint_line.dart';
+export 'src/deriv_chart/chart/worm_chart/worm_chart.dart';
+export 'src/deriv_chart/chart/x_axis/min_candle_duration_for_data_fit.dart';
+export 'src/deriv_chart/deriv_chart.dart';
+export 'src/misc/chart_controller.dart';
+export 'src/models/candle.dart';
+export 'src/models/chart_axis_config.dart';
+export 'src/models/chart_style.dart';
+export 'src/models/tick.dart';
+export 'src/theme/chart_default_dark_theme.dart';
+export 'src/theme/chart_default_light_theme.dart';
+export 'src/theme/chart_theme.dart';
+export 'src/theme/painting_styles/barrier_style.dart';
+export 'src/theme/painting_styles/candle_style.dart';
+export 'src/theme/painting_styles/grid_style.dart';
+export 'src/theme/painting_styles/line_style.dart';
+export 'src/theme/painting_styles/marker_style.dart';
+export 'src/theme/painting_styles/scatter_style.dart';
+export 'src/widgets/market_selector/market_selector.dart';
+export 'src/widgets/market_selector/market_selector_button.dart';
+export 'src/widgets/market_selector/models.dart';
+export 'src/widgets/market_selector/symbol_icon.dart';
